Fix Task 12 review findings: mysqli spans now genuinely nest under root
The real bug behind the "ghost ancestor" spans wasn't a missing
$span->activate() in TracingMiddleware (that was already there since the
first commit, verified via git blame) - it was a TypeError thrown inside
opentelemetry-auto-mysqli's own constructPostHook(). site/config.php and
docker/config.php passed ini_get('mysqli.default_port')'s string return
value straight into `new mysqli(...)`, and MySqliTracker::
storeMySqliAttributes()'s strict `?int $port` parameter type-errors on it.
That TypeError aborts the post-hook before it reaches endSpan(), leaking
the Context scope constructPreHook() pushed - every mysqli_query() on that
connection for the rest of the request then parents to the leaked,
never-exported connect span instead of TracingMiddleware's root. Confirmed
by reproducing via CLI with the real app + real DB, then temporarily
instrumenting a copy of the vendor file inside the container (not
committed - vendor/ is gitignored) to catch the TypeError warning firing
right between the PRE/POST hook log lines.

Fixed at the call site: cast $database_port to (int) in both config.php
files. Verified against a freshly-restarted container: a real curl request
now produces 153 spans (1 root + 20 mysqli::__construct + 132
mysqli_query) that all resolve back to the root span with zero gaps,
confirmed by walking every span's parent chain programmatically via
Jaeger's query API - the Definition of Done is now fully met, not just
"same trace ID" as the original report had to concede.

Also:
- Added a regression test proving TracingMiddleware's activate()/Scope
  mechanism genuinely parents a second span under root, using the SDK
  directly (no native extension needed) - the original 6 tests never had
  two spans in flight at once, so the report's "clean, correctly-parented"
  claim wasn't actually backed by anything.
- Fixed TracingMiddleware::tagFinalStatus() unconditionally overwriting a
  more specific error status description (from its own catch block, or
  from ErrorHandler.php) with a blank one on every 5xx response - guarded
  via the SDK's ReadableSpanInterface::toSpanData() since the API's
  SpanInterface has no getStatus() to check against directly.
- Pinned Dockerfile's `pecl install opentelemetry` to the same 1.2.1
  composer.json's platform pin already claimed, so the two can't silently
  drift apart on the next rebuild.

Full investigation, evidence, and trace stats recorded in
.superpowers/sdd/task-12-report.md's "Review fix round 1" section
(gitignored, not part of this commit).

+++ docker/config.php
@@ -11,7 +11,18 @@ $hostname = getenv('DB_HOST') ?: 'db';
 $username = getenv('DB_USER') ?: 'bcoem';
 $password = getenv('DB_PASSWORD') ?: 'bcoem_password';
 $database = getenv('DB_NAME') ?: 'bcoem';
-$database_port = ini_get('mysqli.default_port');
+// Cast to int (Task 12 review fix): ini_get() always returns a string, and
+// passing that straight to `new mysqli(...)` broke OpenTelemetry's mysqli
+// auto-instrumentation - MySqliTracker::storeMySqliAttributes()'s strict
+// `?int $port` parameter type-errors on a string port, and that TypeError is
+// thrown from INSIDE the extension's post-hook callback, aborting
+// constructPostHook() before it reaches endSpan(). The scope this connect's
+// span pushed onto OTel's ambient Context stack is then never detached -
+// every mysqli_query() span for the rest of the request nests under that
+// leaked, never-exported "ghost" span instead of directly under
+// TracingMiddleware's root span. See task-12-report.md's "Review fix round
+// 1" section for the full investigation.
+$database_port = (int) ini_get('mysqli.default_port');
 
 /**
  * Reuse an existing connection if one is already set (the PHPUnit
