+++ site/config.php
@@ -57,7 +57,19 @@ $database = '';
  * Example: $database_port = 3308;
  */
 
-$database_port = ini_get('mysqli.default_port');
+// Cast to int (Task 12 review fix): ini_get() always returns a string, and
+// passing that straight to `new mysqli(...)` broke OpenTelemetry's mysqli
+// auto-instrumentation - MySqliTracker::storeMySqliAttributes()'s strict
+// `?int $port` parameter type-errors on a string port, and that TypeError is
+// thrown from INSIDE the extension's post-hook callback, aborting
+// constructPostHook() before it reaches endSpan(). The scope this connect's
+// span pushed onto OTel's ambient Context stack is then never detached -
+// every mysqli_query() span for the rest of the request nests under that
+// leaked, never-exported "ghost" span instead of directly under
+// TracingMiddleware's root span. See task-12-report.md's "Review fix round
+// 1" section for the full investigation. Also applies if $database_port is
+// hand-edited to a numeric string like '3308' per the example above.
+$database_port = (int) ini_get('mysqli.default_port');
 
 /**
  * This line strings the information together and connects to MySQL.
